Boxing, enumerações, datas e valores monetários com o BigDecimal

+++ banco/src/modelo/pagamento/Boleto.java
@@ -2,20 +2,22 @@ package modelo.pagamento;
 
 import modelo.Pessoa;
 
+import java.math.BigDecimal;
+
 public class Boleto implements  DocumentoPagavel, DocumentoEstornavel {
     private Pessoa beneficiario;
-    private double valor;
+    private BigDecimal valor;
     private  boolean pago;
 
 
-    public Boleto(Pessoa beneficiario, double valor) {
+    public Boleto(Pessoa beneficiario, BigDecimal valor) {
         this.beneficiario = beneficiario;
         this.valor = valor;
         this.pago = pago;
     }
 
     @Override
-    public double getValorTotal() {
+    public BigDecimal getValorTotal() {
         return valor;
     }
